refactor: Move ErrorBoundary to @components
