+++ packages/client/src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,0 +1,23 @@
+import { Component, ErrorInfo, PropsWithChildren } from 'react';
+
+import { UnexpectedError } from '../../components/UnexpectedError/UnexpectedError';
+
+export class ErrorBoundary extends Component<PropsWithChildren, { hasError: boolean }> {
+  state = { hasError: false };
+
+  static getDerivedStateFromError() {
+    return { hasError: true };
+  }
+
+  componentDidCatch(error: Error, errorInfo: ErrorInfo) {
+    console.error(error, errorInfo);
+  }
+
+  render() {
+    if (this.state.hasError) {
+      return <UnexpectedError />;
+    }
+
+    return this.props.children;
+  }
+}
