+++ packages/client/src/components/ErrorBoundary/index.ts
