--- packages/client/src/providers/ErrorBoundary/types.ts
@@ -1,9 +1,0 @@
-import { ReactNode } from 'react';
-
-export type TErrorBoundaryProps = {
-  children: ReactNode;
-};
-
-export type TErrorBoundaryState = {
-  hasError: boolean;
-};
